More complete intermediate chapter and Rstudio appendix
Need to reread intermediate chapter and Rstudio appendix. Also should consider if keeping the multiple uses of checkout is more appropriate than the new commands the break up it's use. A section on git revert would probably be good including why it is the best action when you have already pushed the commits.