Clean/configure to publish to GitHub pages.
Attempting to render to docs to create a GitHub page. Attempted qurto publish earlier. Neither method has worked but just updated Quarto and RStudio which may help. For example, the command quarto check run in the RStudio terminal was not finding the latest version of quarto prior to updating RStudio.