Draft 2. Added minimal DFG workflow
Draft handles most of the items on my issue checklist and most of the comments in Matt's email. The significant addition was the minimal ADF&G workflow which emphasizes getting the analysis onto GitHub and some rules for our organization account. Also rearranged the intermediate chapter into specific topics with more informative names. The mandate from the preface was moved to Appendix 2 and softened.