Edits associated with second issue on GitHub
New content included additionMost substantial new content about how and when to use the ADFG_DSF account. Some mention about which branches to push and that the network drive should be reference in the readme and kept up to date so that people have access to files in .gitignore. Need to flesh out these suggestions to ensure they are reasonable and feasible both individually and as a group.